Fix audit correctness bugs: CSE stale value, mat*array crash, compiled fallback
Critical/High (with new regression tests): CSE no longer hoists a subexpression across a write to a variable it reads (optimizer.py) -- s*s before and after s=v are different values; mat * array now reports a clean E3402 instead of crashing with AttributeError (type_checker.py).

Compiled path: removed the worker-thread torch._dynamo.reset() on cache eviction and reverted the fallback to pristine bindings (compiled.py); completed the scalar-loop pow base-clamp fix (codegen.py).

Low: prepare_output int-dtype mean via .float(); scalar %-by-zero epsilon; hash_int [0,max_val) range; len() rejects non-string/array; clear 'const arrays unsupported' error. All 1373 tests pass; no perf regression (back-to-back benchmark confirmed). Scatter += x B*H*W left unchanged -- intended per-source splat semantics (examples/caustics.tex).

+++ tex_marshalling.py
@@ -245,7 +245,7 @@ def prepare_output(raw: torch.Tensor | str, output_type: str) -> Any:
     elif output_type == "FLOAT":
         if raw.dim() == 0:
             return raw.item()
-        return raw.mean().item()
+        return raw.float().mean().item()  # .float(): mean() rejects integer dtypes
 
     elif output_type == "LATENT":
         # LATENT expects [B, C, H, W] — permute back from channel-last
@@ -262,6 +262,6 @@ def prepare_output(raw: torch.Tensor | str, output_type: str) -> Any:
     elif output_type == "INT":
         if raw.dim() == 0:
             return int(raw.item())
-        return int(raw.mean().item())
+        return int(raw.float().mean().item())  # .float(): mean() rejects integer dtypes
 
     return raw.cpu()
